refactor: mv `closableStringBuilder` from `svc_override.go` to `override.go`
