+++ internal/pkg/cli/override.go
@@ -3,6 +3,12 @@
 
 package cli
 
+import (
+	"fmt"
+	"io"
+	"strings"
+)
+
 const (
 	// IaC options for overrides.
 	cdkIaCTool = "cdk"
@@ -17,4 +23,18 @@ var validIaCTools = []string{
 
 var validCDKLangs = []string{
 	typescriptCDKLang,
+}
+
+type stringWriteCloser interface {
+	fmt.Stringer
+	io.WriteCloser
+}
+
+type closableStringBuilder struct {
+	*strings.Builder
+}
+
+// Close implements the io.Closer interface for a strings.Builder and is a no-op.
+func (sb *closableStringBuilder) Close() error {
+	return nil
 }
